Rest Api completed with curd operationt

+++ accounts/src/main/java/com/dash/accounts/dto/AccountsDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AccountsDto {
-    private int account_number;
+    private Long account_number;
     private String account_type;
     private String branch_address;
 }
